7
First sound clear ArrayList instance objects to create a method called createAppointment to create a new booking, to create a method called printExistingAppointments, used to display in the ArrayList existing reservation. Create a method called cancelBooking for cancelling appointments using a patient's phone. If the phone is not found in an existing appointment, print a message to eliminate this error